+++ src/main/java/com/blu/curator/ExecutesSecondsClient.java
@@ -23,6 +23,6 @@ public class ExecutesSecondsClient {
         // second one
         SimpleClient secondClient = new SimpleClient("second", client, PATH);
         secondClient.start();
-        Thread.sleep(Integer.MAX_VALUE);
+        Thread.sleep(Long.MAX_VALUE);
     }
 }
